Update README.md
Updates in implementing our own persister strategy. Added json_encode for 'changed' and 'meta' fields as they are strings and weren't being saved without encoding them. Also added the 'original' field which tells the original state of the data before it was changed. P.S. there was a little typo.